+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/Mage.java
@@ -663,4 +663,6 @@ public interface Mage extends CostReducer, CooldownReducer {
     MageEffectContext getEffectContext();
 
     long getCreatedTime();
+
+    boolean trigger(String trigger);
 }